requirements.txt
dash == 1.20.0
dash-bootstrap-components == 0.13.0
dash-core-components == 1.16.0
dash-html-components == 1.1.3
plotly == 4.14.3
pandas == 1.2.4
numpy == 1.19.5
dash-table == 4.11.3
PyYAML == 5.4.1
psycopg2 == 2.8.6